atomized sms module, updated refs and tests
