+++ src/controller/sms/get_step_name.js
@@ -1,0 +1,18 @@
+const steps = require('./steps')
+const debug = require('debug')('sms') // eslint-disable-line
+
+/**
+ * getNextStepName - gets the current step name based on session data
+ *
+ * @param  {object} session session object from request
+ * @return {string} name of current step
+ */
+module.exports = function getStepName(session) {
+  const stepName = Object.keys(steps).reduce((memo, step) => {
+    if (memo) return memo
+    if (!session[step]) return step
+    return null
+  }, null)
+
+  return stepName
+}
