+++ src/controller/sms/message.js
@@ -1,0 +1,10 @@
+// TWIML markup helper
+module.exports = function message(msg) {
+  return `
+    <Response>
+      <Message>
+        ${msg}
+      </Message>
+    </Response>
+  `
+}
