+++ src/controller/sms/sms.js
@@ -1,0 +1,5 @@
+const dispatcher = require('./dispatcher')
+
+module.exports = {
+  dispatcher,
+}
